Reduce memory footprint of Url instances
Was 160 bytes, now 64 bytes.

+++ core/include/tangram/util/url.h
@@ -118,7 +118,7 @@ private:
     // parts describes URL components by their location within the buffer.
     struct Parts {
         struct Range {
-            size_t start = 0, count = 0;
+            uint16_t start = 0, count = 0;
         } scheme, location, path, parameters, query, fragment, media, data;
     } parts;
